requirements.txt
alembic==1.4.3
attrs==20.2.0
cffi==1.14.3
click==7.1.2
cryptography==2.9.2
flasgger==0.9.5
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwt==1.0.0
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.15.0
python-editor==1.0.4
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
xlrd==1.2.0
